просто тест

+++ domain/models/orderResponse.go
@@ -3,6 +3,6 @@ package models
 type OrderResponse struct {
 	Number     string  `json:"number"`
 	Status     string  `json:"status"`
-	Accrual    float32 `json:"accrual"`
+	Accrual    float64 `json:"accrual"`
 	UploadedAt string  `json:"uploaded_at"`
 }
